airflow/requirements.txt
google-cloud-storage
requests
dbt-core
dbt-bigquery